requirements.txt
Flask>=2.3
requests>=2.31
python-dotenv>=0.21